Move socket extensions to own module
